python/translation/pyproject.toml
[tool.poetry]
name = "translation"
version = "0.1.0"
description = ""
authors = ["Jaz <ericvolp12@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.10"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
pydantic = "^1.10.7"
transformers = "^4.45.1"
numpy = "^1.24.3"
scipy = "^1.10.1"
python-json-logger = "^2.0.7"
prometheus-fastapi-instrumentator = "^6.0.0"
opentelemetry-api = "^1.17.0"
opentelemetry-sdk = "^1.17.0"
opentelemetry-instrumentation-fastapi = "^0.43b0"
opentelemetry-exporter-otlp-proto-http = "^1.17.0"
timm = "^0.9.2"
optimum = "^1.16.2"
asyncio = "^3.4.3"
opentelemetry-instrumentation-aiohttp-client = "^0.43b0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
